requirements.txt
requests
python-dotenv
csv